Refactor unittest and add new test checking that dice rolls differ
InputTypes, a list of all valid input types, was used in two different
tests but was not shared. The refactoring moved it up a scope and
allowed it to be shared instead of duplicated.

A test to check that dice rolls didn't always return the same was
added. This was something that was missing previously and might be
useful to assure the validity of future changes.