Change randomDie rng to value type
This change makes randomDie consistent with other random ranges w.r.t.
storing the rng as a value instead of as a pointer. This will prevent
issues arising in code such as this:

    auto foo()
    {
        Random rng;  // will be deleted on exit of this scope
        return randomDie([0.5, 0.5], rng); // using later will error
    }

Additionally, the documentation has been updated to display a warning
similar to the warnings given to randomCover and randomSample to alert
users of the common pitfall of using value-based rngs.

Finally, a test assuring that using the default rng is safe from this
pitfall has been added.