Expand scope of unittests of RandomDie and fix immutable bug
Previously, a bug concerning immutable reals slipped through the
cracks. The unittests now test against const and immutable types to
cover those kinds of cases. Example of code that would fail to compile:

    immutable real[] props = [0.0, 1.0];
    auto dicer = randomDie(props);

To make the unittest pass, a simple fix was also implemented in the
way that the StorageType is chosen by default.