Change the way RandomDie handles qualified types
This makes it so that randomDie explicitly supports Ranges that have
qualified ElementTypes (e.g. immutable, const, shared). In addition it
changes the way that RandomDie stores the cumulative array internally
to be immutable (since it will never change after being made).

This change would make it so that qualified StorageTypes are equivalent
to their unqualified forms, so to make that clear only unqualified
StorageTypes are accepted. The documentation and unittests for
RandomDie have been updated to reflect this change.