Fix issue 11597 - Binary search for std.random.dice
This commit adds a new overload for std.random.dice that enables it
to utilize binary search to improve performance when a cumulative sum
SortedRange is passed in. This change has the potential to vastly
increase performance when it is applicable and has a relatively low
chance of impacting existing code.

The solution does not prefer a particular sorting order; it supports
either less than or greater than orderings equally (it must still be
a cumulative sum array, but it can be done in either direction).

Documentation is included to exhibit the enhancement.