Improved unit tests and fixed compilation issue
`float.min` has been deprecated in favor of `float.min_normal`. This
was missed before, so the unit tests have been updated to make sure
to cover at least one element from each major type of numeric. So
it now covers an integral, an unsigned, and a floating point type.

In addition, a suggested unit test was added to test the new feature
against the equivalent call using the original way.