Avoid calling a qualified constructor in Cycle.
This is actually invalid code, as typeof(this) == inout(this),
but Cycle has no inout constructor. This wasn't previously
detected, as the constructor was mistakenly flagged as creating
an unique object where qualifiers can be disregarded. (The
reference parameter is leaked into the Cycle instance, so
this is obviously not the case.)