Fix issue# 12507: SysTime.init.toString() blows up.
It didn't used to be possible for SysTime.init to have a non-null
timezone member, since member variables of class type did not used to be
able to initialized at compile time. That is no longer the case, and in
order to make it so that toString returns a special value, SysTime.init
now has a special timezone which is unique to it which makes it so that
it's treated kind of like NaN in that until you assign a value to it, it
retains the same value regardless of the operation called on it.

SysTime.init is now equivalent to 0001-01-01T00:00:00+00:00.