Add doc reference to std.uni, std.array
In a recent debate on forums there were some complaints the some functions to perform common string operations did not appear in std.string, but in std.uni or std.array.  Which made them hard to find. This update simply points readers to those libraries, in the event that they don't find the function they wanted in std.string.