Add support of IeeFlags and FloatingPointControl for PPC32/64.
This corrects some existings constants and adds new constants and logic.
Reading the PowerPC documentation is sometimes confusing because not only the
CPU is big endian but also the bit numbering inside a word is reversed (old
IBM style big endian).
The functionaly has been verified on a Linux/PPC64le system with LDC. The LDC
specific implementation is left out.