Add Params/Returns ddoc to free functions in std.random
The existing docs have not been tweaked in any way, so the new additions
result in some duplication of information.  This seems a better first
step than a more major revision.

Documentation for struct methods has been left unchanged, and will be
addressed in a later patch.

I've also left the documentation for uniformDistribution unchanged as
IMHO this function is so broken in its design, it really ought to be
deprecated and removed :-)