Fix Issue 5036 - Remove caching from Sequence
This appears to be the last caching range in Phobos.

This is a potentially breaking change: we lose the ability to moveFront, and obviously there may be performance implications if the function is expensive and users are calling front more than once. I think this is still a worthwhile change since it is better to have a uniform policy of no caching on ranges rather than arbitrary selection of ranges that do caching.

https://issues.dlang.org/show_bug.cgi?id=5036