Improve docs for until().
Removed docs from struct Until, because it serves no purpose except to
clutter the docs with another obtuse blob of text in the overload list,
yet none of its members are even documented to begin with. It should
suffice for the docs of until() to describe what can be done with the
returned range.