fix Issue 14005 - core.exception.InvalidMemoryOperationError@(0) during byLine
The problem was that readlnImpl assumed that it's safe to write beyond buf
when the GC block is large enough. That's not true. The space may already
be in use otherwise. In the original test case readlnImpl would write over
internal data of the GC at the end of the block.