Tuple's toString made useful.
http://forum.dlang.org/thread/srofuqyrdzphgzylbwds@forum.dlang.org
Format Tuple's fields with format in %(%|%), similar to arrays, or, without %|, each filed specifically.

Doesn't break any existing code as %s qualifier's behavior is not modified

The overload without parameters is intentionally left as a template, as it appears
to reveal some bugs in std.format for certain types. Remove the
template parameter list to reproduce.