Update to logger due to fracSec deprecation in std.datetime.
I also fixed it so that it casts to DateTime rather than calling all of
the various properties on SysTime individually, since that's more or
less equivalent to casting to DateTime and getting the property off of
it for each call. I probably never should have put those properties on
SysTime in the first place, though it is more user-friendly in the cases
where you don't care about efficiency. I expect that we _do_ care about
efficiency in the logger however.