Port std.zip tests from the dmd test suite to std.zip.
For whatever reason, the dmd test suite has tests for std.zip, though
for the most part, they seem to just run it without actually checking
the results, making them of questionable value. But this commit ports
them to std.zip (to remove a Phobos dependency in the dmd test suite)
and attempts to make it so that they actually check the results. So, the
tests are not identical by any means, but they're similar. Perhaps the
most significant test is the one that tests a file that's written with
std.zip against the unzip utility, since that's testing against
something that we didn't write.