Fix missing ddoc'd unittests for {does,may}PointTo.
Apparently private function was inadvertently inserted between the
original pointsTo implementation and its ddoc'd unittests, causing ddoc
not to pick them up correctly. So reordered the private function and
deprecated alias so that the unittests immediately follow the pointsTo
overloads.

Also, did some cosmetic fixups in the unittests to make them nicer to
read as docs.