New primitive - Tiny UTF Binary Search Table
This lays on the foundation of UTF word notion,
a 32-bit integer that contains UTF encoded codepoint.
Top non-zero byte is the UTF-8 starter, so it's BigEndian
when reading directly from strings.

Same applies to UTF-16 word.

Also this invents simpler API for matching that easily combines:
 - length of code point matched
 - result of match (true/false if belongs to this set)
 - bad encoding flag

in one packed 32-bit machine word - UtfLookup. This struct is
going to be reused for all of "directly on UTF" matchers in std.uni.