removed template-parens in non-template function
commented out for auto-test; to investigate SIGSEGV

Uncomment code

Fix infinite recursion in nullary Tuple.toString

This overload is intentionally left as a template, as it appears
to reveal some bugs in std.format for certain types. Remove the
template parameter list to reproduce.

Fix Tuple formatting of shared fields