fix documentation for symlink
The target path is actually relative to the link, not to the current
working directory. This is in line with the ln shell command, so changing
the documentation to match the behavior seems in order.

Also changing some wording, and un-highlighting "symlink" throughout as
it never refers to the function.