Implement std.array.staticArray.
staticArray makes it easier to create static arrays in-place (without a
separate declaration).

One overload is used to convert a range into a static array.
This requires the caller to specify the size as a compile-time argument.

Examples:
    int[3] = only(1,2,3).staticArray!3   // [ 1, 2, 3 ]
    int[4] = only(1,2,3).staticArray!4   // [ 1, 2, 3, 0 ]
    int[3] = only(1,2,3,4).staticArray!3 // [ 1, 2, 3 ]

As shown, staticArray!n discards elements of the range beyond `n`, and
leaves as default elements that the source range is not long enough to
fill.

The second overload converts an array literal into a static array,
compensating for the language's lack of a static array literal syntax.

Example:
    auto a = [1, 2, 3].staticArray(1,2,3); // a is an int[3]

This is preferable to

int[3] a = [1,2,3];

as you don't have to determine the count manually (and update it each
time you change the array literal).

staticArray supports immutable elements by using emplaceRef with an
Unqual!E[n] that is converted to the desired type on return.