Avoid brute force fd close() on Linux
The current Posix implementation of spawnProcess calls close on
every fd from 3 to the max fd count (as gathered by rlimit()).
Given that the process doesn't have that many descriptors open
in the *vast* majority of cases, this is quite pessimal.

Instead, use poll() to see which descriptors are actually open,
and close those.