Add example of stringLambda with generated names.
This example shows how aliasSeqOf can be used to generate names when
using stringLambda with a many-parameter function.

The example is a bit convoluted as the chars of std.ascii.lowercase
must be mapped to individual strings.

An alternative is to allow paramNames to take chars as well as strings,
but this would complicate the implementation.