Implement post overload in std.net.curl
This overload provides the ability to send properly escaped
www-form-urlencoded data using an associative array. The data is built into
name value pairs using an Appender!string that reserves 128 bytes per name
value pair. An example was added to the documentation and the existing one
changed as this is the more common use case.

Several corrections to coding style were suggested and corrected.